// artic-prime-react/src/types.ts
export interface Artwork {
  id: number
  title?: string
  place_of_origin?: string
  artist_display?: string
  inscriptions?: string
  date_start?: number
  date_end?: number
}

export interface ArtworksApiResponse {
  data: Artwork[]
  pagination?: {
    total?: number
    limit?: number
    offset?: number
    total_pages?: number
    current_page?: number
  }
}